Move server code into a subpackage
